Remove EPSILON checks in 1D gf's.
If the random number to drawTime is smaller than EPSILON that doesn't necesserily mean that t = 0. What if D is for example very small. Same for when r0 is very close (EPSILON*L) to a boundary, that doens't mean it necesseraly escapes.

Found error because of missing check gf.drawTime( 1e-16 ) for 1D Rad.